WIP: Refactoring to more general offline plotting library
