+++ source/plot_types/bar.cpp
@@ -1,0 +1,29 @@
+/*
+ * bar.cpp Implementation of BarPlot class methods and operators
+ *
+ * References:
+ *
+ *   [1]
+ *
+ * Future Improvements:
+ *
+ *   [1] 
+ *
+ * Author:              Tom Clark  (thclark@github)
+ *
+ * Copyright (c) 2017-8 T Clark. All Rights Reserved.
+ *
+ */
+
+#include <json/single_include/nlohmann/json.hpp>
+#include "bar.h"
+#include "figures.h"
+
+
+using nlohmann::json;
+
+
+namespace plotly {
+
+
+} // end namespace plotly
